+++ Sources/QuillCodeApp/QuillCodeComposerSurface.swift
@@ -9,9 +9,16 @@ public struct ComposerSurface: Codable, Sendable, Hashable {
     public var canSend: Bool
     public var slashSuggestions: [SlashCommandSuggestionSurface]
     public var fileMentionSuggestions: [FileMentionSuggestionSurface]
+    /// Previously sent user messages (oldest first) for Up/Down history recall.
     public var sentMessageHistory: [String]
+    /// Bumped by the `focus-composer` command; the view focuses the input when it changes.
     public var focusToken: Int
+    /// The current run's plan progress for the always-visible strip above the input. nil ⇒ no plan ⇒
+    /// the strip renders nothing (a plan-less session looks byte-identical). Optional, so it decodes
+    /// safely from surfaces persisted before this field existed.
     public var planProgress: WorkspacePlanProgress?
+    /// Composer submissions entered during the live run, shown as delete-able chips above the
+    /// input and drained one per turn boundary. Empty when there is nothing queued.
     public var followUpQueue: [FollowUpItemSurface]
 
     public init(
@@ -25,6 +32,9 @@ public struct ComposerSurface: Codable, Sendable, Hashable {
         self.draft = composer.draft
         self.placeholder = composer.placeholder
         self.isSending = composer.isSending
+        // Sendable even while a run is in flight: a non-empty draft can always be submitted —
+        // it enqueues as a follow-up chip when sending, and sends immediately when idle. The
+        // composer never locks, so `canSend` no longer gates on `isSending`.
         self.canSend = !composer.draft.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
         self.slashSuggestions = SlashCommandCatalog.suggestions(for: composer.draft)
         self.fileMentionSuggestions = FileMentionCatalog.suggestions(
@@ -57,6 +67,7 @@ public struct ComposerSurface: Codable, Sendable, Hashable {
         )
         self.sentMessageHistory = try container.decode([String].self, forKey: .sentMessageHistory)
         self.focusToken = try container.decode(Int.self, forKey: .focusToken)
+        // Both optional-add fields decode safely from surfaces persisted before they existed.
         self.planProgress = try container.decodeIfPresent(WorkspacePlanProgress.self, forKey: .planProgress)
         self.followUpQueue = try container.decodeIfPresent([FollowUpItemSurface].self, forKey: .followUpQueue) ?? []
     }
@@ -76,6 +87,8 @@ public struct ComposerSurface: Codable, Sendable, Hashable {
     }
 }
 
+/// A single queued follow-up rendered as a composer chip. Carries the id (for the delete
+/// affordance's target) and the text (the chip label / accessibility).
 public struct FollowUpItemSurface: Codable, Sendable, Hashable, Identifiable {
     public var id: UUID
     public var text: String
