+++ Sources/QuillCodeApp/WorkspaceModelSlashCommands.swift
@@ -1,48 +1,13 @@
 import Foundation
-import QuillCodeAgent
 import QuillCodeCore
 
 @MainActor
 extension QuillCodeWorkspaceModel {
-    func drainSelfHealingNotices(expectedThreadID: UUID?) {
-        guard let channel = retryEventChannel else { return }
-        let events = channel.drain()
-        guard !events.isEmpty,
-              let thread = selectedThread,
-              thread.id == expectedThreadID
-        else { return }
-        mutateSelectedThread { thread in
-            for event in events {
-                thread.events.append(ThreadEvent(
-                    kind: .notice,
-                    summary: SelfHealingNoticePlanner.noticeSummary(attempt: event.attempt, kind: event.kind)
-                ))
-            }
-        }
-        if let thread = selectedThread {
-            threadPersistence.save(thread)
-        }
-    }
-
-    func executeBrowserToolForAgent(_ call: ToolCall, workspaceRoot: URL) -> ToolResult? {
-        let result = mutateBrowserState { browser, lastError in
-            WorkspaceBrowserToolExecutor.execute(
-                call,
-                workspaceRoot: workspaceRoot,
-                browser: &browser,
-                lastError: &lastError,
-                domainPolicy: root.config.browserDomainPolicy
-            )
-        }
-        refreshTopBar(agentStatus: root.topBar.agentStatus)
-        return result
-    }
-
     func handleSlashCommand(_ command: SlashCommand, originalPrompt: String, workspaceRoot: URL) async {
         let action = WorkspaceSlashCommandDispatchPlanner.action(
             for: command,
             userText: originalPrompt,
-            statusText: composerStatusText()
+            statusText: statusText()
         )
         await runSlashCommandDispatchAction(action, workspaceRoot: workspaceRoot)
         composer.isSending = false
@@ -108,7 +73,7 @@ extension QuillCodeWorkspaceModel {
         appendLocalCommandTranscript(transcript)
     }
 
-    private func composerStatusText() -> String {
+    private func statusText() -> String {
         WorkspaceStatusTextBuilder.statusText(for: WorkspaceStatusContextBuilder.context(
             root: root,
             selectedProject: selectedProject,
